Added a function that changes the user's direction so they're always facing the direction of the room they entered
Wrote a function that changes the user's direction when they enter a room. If they enter a room to the left, then the user will now be facing left. Also added to the description string, which now tells the user which direction they are facing and what directions the available pathways are in.